refactor: folder structure refactor/changed to feature based structure
