+++ src/features/3d-presentation/components/box.tsx
