+++ src/features/3d-presentation/components/scroll-scene.tsx
@@ -13,8 +13,8 @@ export function Scene({ currentModelIndex, models }: SceneProps) {
       style={{ background: "transparent" }}
     >
       <ambientLight intensity={0.4} />
-      <pointLight position={[10, 10, 10]} intensity={0.8} />
-      <pointLight position={[-10, -10, -10]} intensity={0.3} />
+      <pointLight position={[3, 3, 3]} intensity={0.8} />
+      <pointLight position={[-3, -3, -3]} intensity={0.3} />
       {models?.[currentModelIndex]}
     </Canvas>
   );
